Fixes #6 whitelist not found when using sudo
This fixes #6 by changing the location of the whitelist.txt file.
Instead of storing it in the current users home folder, it should be
created in /etc/pihole.  The script will also create the directory if
it does not exist, but will not add a whitelist.txt by default.  You
would still need to go back and manually add one.  But this directory
will be used to store other support files as the script improves.

I also added sudo commands to the commands requiring elevation.

Finally, changed the comment to reflect a more accurate number of
domains getting blocked.