experimenting with more accurate domain list
wormhole.sh will attempt to remove domains that no longer exist by
running the dig command on each one and seeing if there is an answer
section.  If it doesn't have it, it is not added to the list.

This script is incredibly slow (even on a Mac), so you may want to just
run it overnight or it might even take a couple days depending on the
size of your list.

I also uploaded a pihole movie file, which I am using for blocking Hulu
video ads.