Testing script for a mini LCD
This script is for use with an Adafruit LCD so that you can peek at the
domains as they are blocked in real time.  It's not necessary for the
Pi-hole to function, but something fun I am working on.  There will be
a video in the next article I post.