Added nginx as an optional server.
Why? For the glory of satan of course...
Also, many people tends to use their PIs for multiple purposes. As mail
servers, webpage hosting etc. Usually in that cases nginx is used as far
more powerful alternative to lighttpd, and fair more lighter than
apache. This patch allows them to unleash the power of pi-hole without
too much hacking.

Also, I'm the one of those people, and needed it ;).