Remove hardcoded local-ttl value from dnsmasq config
- The local ttl value tells clients to cache local responses for 5 mins.
  This makes it so /etc/host and dhcp updates can take up to 5 mins to
propigate. In almost all cases these should be instant.

From DNSMasq man page:

When replying with information from /etc/hosts or configuration or the
DHCP leases file dnsmasq by default sets the time-to-live field to zero,
meaning that the requester should not itself cache the information. This
is the correct thing to do in almost all situations. This option allows
a time-to-live (in seconds) to be given for these replies. This will
reduce the load on the server at the expense of clients using stale data
under some circumstances.