Do not ask admin password just for help documentation
It is not a good idea to ask for an admin password just to get the
command documentation.
"pihole -h" should return the arguments documentation with no need to
enter a password.

Without the patch I get:
$ ./pihole -h
Password:

It is easy to reproduce by invalidating the sudo password using
"sudo -k" before running pihole again.

Signed-off-by: Ludovic Rousseau <ludovic.rousseau@free.fr>