Fix "bad address in /etc/pihole/gravity.list" in pihole.log
This puts gravity.list in the proper format and
prefixs the IP address of the pi to the beginning
of each line. This allows dnsmasq to properly parse
the file, and actually use the content of gravity.list.

This fixes #758, but works any time the list is built,
not just on install (Example: `pihole -g`)

Signed-off-by: Andrew <agh1467@rcpt.at>