Add sed command to replace crlf if they exist in setupVars
This could be added before "source ${setupVars}" (line 2548) or in the "accountForRefactor" function.  If "accountForRefactor" will go away in the future, I would leave it where it is.
This could save hours of troubleshooting when the setupVars has crlf.  The symptoms are a partial failure.  The only functions that fail are Boolean matches, for example "INSTALL_WEB_SERVER=true" does not match true with a crlf after it.  Collecting other information, like IP addresses and domains, will still work even followed by a crlf.