Update index.php
Why? When forwarding the web interface of pihole through ssh tunnel to localhost:xx I didn't get the splash page. With this change I can access the splash page correctly and don't get told, that localhost was blocked because "it wasn't found on any blacklist". Whitelisting localhost did not work, but this small change does the trick.