Consequently use $setupVars variable
In two cases, the /etc/pihole/setupVars.conf path was used literally rather then the defined $setupVars variable. While this is not a real issue, this change makes it easier to change the paths, when required at some time in the future.

Additionally all near assigned strings are quoted, which is not necessary here but aligns with the coding standard of the script and highlights the strings as such in most editors and development platforms for developer convenience.

Signed-off-by: MichaIng <micha@dietpi.com>