Discourage inexperienced users to purge dependencies
For reference: https://github.com/pi-hole/pi-hole/issues/3929

The ability to loop through packages for removal, and the statement "All dependencies are safe to remove on Raspbian" has been added at a time where a much smaller list of dependencies were installed, all indeed relatively safe to purge. Nowadays this list has grown and includes important system packages, like iproute2, psmisc, sudo, curl and others, which are often again dependencies of other packages, like network stacks (ifupdown) and others, so that inexperienced users, following that statement, may break their systems network capabilities and more.

The outdated info message has hence been removed, replaced with a recommendation to not remove packages, and the default answer, whether to loop through packages or not, has been changed to "No" accordingly. When "Yes" was still selected, a more prominent "yellow" message is printed to repeat that recommendation for individual packages as well, when the user is uncertain.

Signed-off-by: MichaIng <micha@dietpi.com>