Install on IPv6-only/DNS64/NAT64 system
Alters the test for IPv4 route so the installer doesn't fail instantly on IPv6

If a valid IPv6 address is found, but no IPv4, runs connectivity check to github

If all appears well gives user a dialog to proceed with IPv6 only installation

Switches DNS server options to only contain IPv6 addresses.

If DNS64 and NAT64 are not installed on the system, it will fail gracefully at
the github connectivity test.

Adds new environment variable PIHOLE_IPV6_ONLY to bypass dialog and tests.

Signed-off-by: Rob Gill <rrobgill@protonmail.com>